rpios file commited

+++ include/utils.h
@@ -9,12 +9,12 @@
  * Used for reading/writing hardware registers safely.
  */
 
-static inline void mmio_write(uint32_t reg, uint32_t value)
+static inline void mmio_write(uintptr_t reg, uint32_t value)
 {
     *(volatile uint32_t *)reg = value;
 }
 
-static inline uint32_t mmio_read(uint32_t reg)
+static inline uint32_t mmio_read(uintptr_t reg)
 {
     return *(volatile uint32_t *)reg;
 }
